% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensing.R
\name{ssdtests_licensing_md}
\alias{ssdtests_licensing_md}
\title{Licensing Markdown}
\usage{
ssdtests_licensing_md()
}
\description{
A string of markdown code indicating the licensing of the code
and documentation
}
\examples{
ssdtests_licensing_md()
}
