% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdtests-package.R
\docType{package}
\name{ssdtests-package}
\alias{ssdtests}
\alias{ssdtests-package}
\title{ssdtests: Tests for ssdtools}
\description{
Tests for ssdtools package.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au}
  \item David Fox \email{david.fox@environmetrics.net.au}
  \item Carl Schwarz
}

Other contributors:
\itemize{
  \item Angeline Tillmanns [contributor]
  \item Seb Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Kathleen McTavish [contributor]
  \item Doug Spry [contributor]
  \item Rick van Dam [contributor]
  \item Graham Batley [contributor]
  \item Ali Azizishirazi [contributor]
  \item Province of British Columbia [funder, copyright holder]
  \item Environment and Climate Change Canada [funder, copyright holder]
  \item Australian Government Department of Climate Change, Energy, the Environment and Water [funder, copyright holder]
}

}
\keyword{internal}
